+++ examples/react/src/App.tsx
@@ -7,6 +7,7 @@ function App() {
       <OverlayScrollbarsComponent
         style={{ width: '222px', height: '222px' }}
         options={{ scrollbars: { theme: 'os-theme-light' } }}
+        defer
       >
         <img src={logo} className="App-logo" alt="React logo" width="333" height="333" />
       </OverlayScrollbarsComponent>