+++ packages/overlayscrollbars-vue/src/useOverlayScrollbars.ts
@@ -1,4 +1,4 @@
-import { shallowRef, unref, watch } from 'vue';
+import { onUnmounted, shallowRef, unref, watch } from 'vue';
 import { OverlayScrollbars } from 'overlayscrollbars';
 import type { Ref, UnwrapRef } from 'vue';
 import type { InitializationTarget } from 'overlayscrollbars';
@@ -6,6 +6,11 @@ import type {
   OverlayScrollbarsComponentProps,
   OverlayScrollbarsComponentRef,
 } from './OverlayScrollbarsComponent.types';
+
+type Defer = [
+  requestDefer: (callback: () => any, options?: OverlayScrollbarsComponentProps['defer']) => void,
+  cancelDefer: () => void
+];
 
 export interface UseOverlayScrollbarsParams {
   /** OverlayScrollbars options. */
@@ -16,15 +21,56 @@ export interface UseOverlayScrollbarsParams {
   events?:
     | OverlayScrollbarsComponentProps['events']
     | Ref<OverlayScrollbarsComponentProps['events']>;
+  /** Whether to defer the initialization to a point in time when the browser is idle. (or to the next frame if `window.requestIdleCallback` is not supported) */
+  defer?: OverlayScrollbarsComponentProps['defer'] | Ref<OverlayScrollbarsComponentProps['defer']>;
 }
 
-export type UseOverlayScrollbarsInitialization = (
-  target: InitializationTarget
-) => OverlayScrollbars;
+export type UseOverlayScrollbarsInitialization = (target: InitializationTarget) => void;
 
 export type UseOverlayScrollbarsInstance = () => ReturnType<
   OverlayScrollbarsComponentRef['osInstance']
 >;
+
+const createDefer = (): Defer => {
+  /* c8 ignore start */
+  if (typeof window === 'undefined') {
+    // mock ssr calls with "noop"
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
+    const noop = () => {};
+    return [noop, noop];
+  }
+  /* c8 ignore end */
+
+  let idleId: number;
+  let rafId: number;
+  const wnd = window;
+  const idleSupported = typeof wnd.requestIdleCallback === 'function';
+  const rAF = wnd.requestAnimationFrame;
+  const cAF = wnd.cancelAnimationFrame;
+  const rIdle = idleSupported ? wnd.requestIdleCallback : rAF;
+  const cIdle = idleSupported ? wnd.cancelIdleCallback : cAF;
+  const clear = () => {
+    cIdle(idleId);
+    cAF(rafId);
+  };
+
+  return [
+    (callback, options) => {
+      clear();
+      idleId = rIdle(
+        idleSupported
+          ? () => {
+              clear();
+              // inside idle its best practice to use rAF to change DOM for best performance
+              rafId = rAF(callback);
+            }
+          : callback,
+        typeof options === 'object' ? options : { timeout: 2233 }
+      );
+    },
+    clear,
+  ];
+};
 
 /**
  * Composable for advanced usage of OverlayScrollbars. (When the OverlayScrollbarsComponent is not enough)
@@ -39,7 +85,17 @@ export const useOverlayScrollbars = (
   let instance: ReturnType<UseOverlayScrollbarsInstance> = null;
   let options: UnwrapRef<UseOverlayScrollbarsParams['options']>;
   let events: UnwrapRef<UseOverlayScrollbarsParams['events']>;
+  let defer: UnwrapRef<UseOverlayScrollbarsParams['defer']>;
   const paramsRef = shallowRef(params || {});
+  const [requestDefer, clearDefer] = createDefer();
+
+  watch(
+    () => unref(paramsRef.value?.defer),
+    (currDefer) => {
+      defer = currDefer;
+    },
+    { deep: true, immediate: true }
+  );
 
   watch(
     () => unref(paramsRef.value?.options),
@@ -69,14 +125,25 @@ export const useOverlayScrollbars = (
     { deep: true, immediate: true }
   );
 
+  onUnmounted(() => {
+    clearDefer();
+    instance?.destroy();
+  });
+
   return [
-    (target: InitializationTarget): OverlayScrollbars => {
-      // if already initialized return the current instance
+    (target) => {
+      // if already initialized do nothing
       if (OverlayScrollbars.valid(instance)) {
         return instance;
       }
 
-      return (instance = OverlayScrollbars(target, options || {}, events || {}));
+      const init = () => (instance = OverlayScrollbars(target, options || {}, events || {}));
+
+      if (defer) {
+        requestDefer(init, defer);
+      } else {
+        init();
+      }
     },
     () => instance,
   ];